Clean up xml_native_schema.cc.
Four local refactors to the MJCF XSD generator; all preserve the schema
semantics (same validating behaviour, same set of declared types).

1. Move the ~600-line kMjXAttrTable (and the mjXAttr/mjXAttrKind type
   declarations) into xml_native_reader.cc / xml_native_reader.h, next
   to MJCF[] and the enum maps. Reverts the earlier `extern const mjMap`
   sprinkling: the maps go back to internal linkage, and the schema
   generator just consumes kMjXAttrTable through its declarations.
   xml_native_schema.cc shrinks from 1653 to 532 lines.

2. Drop the manual kGeomTypesSz / kNDYN / etc. shadow constants; reference
   the authoritative mjNGEOMTYPES / mjNDYN / ... enum values directly from
   mjmodel.h / user_composite.h / user_flexcomp.h. No more silent drift.

3. Replace the O(N) linear Lookup() over kMjXAttrTable with a file-static
   unordered_map<(string_view, string_view), const mjXAttr*>, populated
   on first call. Schema emission drops from O(N^2) to O(N).

4. Replace the ListTypeName / EmitListType string round-tripping with a
   ListType struct (kind + size). EmitListType takes the struct directly,
   no stoi/find_last_of parsing of type names.

5. Deduplicate <xs:simpleType> enum declarations by mjMap* identity: the
   first (element, attr) pair that references a given map sets the
   canonical enum type name; every subsequent attribute using the same
   map reuses it. The generated XSD shrinks from 183 enum simpleTypes
   (many byte-identical copies of bool / interp / enable enums) down to
   39 distinct ones.
