+++ src/xml/xml_native_reader.cc
@@ -602,42 +602,42 @@ std::vector<const char*> MJCF[nMJCF] = {
 //---------------------------------- MJCF keywords used in attributes ------------------------------
 
 // coordinate type
-extern const mjMap coordinate_map[2] = {
+const mjMap coordinate_map[2] = {
   {"local",   0},
   {"global",  1}
 };
 
 
 // angle type
-extern const mjMap angle_map[2] = {
+const mjMap angle_map[2] = {
   {"radian",  0},
   {"degree",  1}
 };
 
 
 // bool type
-extern const mjMap bool_map[2] = {
+const mjMap bool_map[2] = {
   {"false",   0},
   {"true",    1}
 };
 
 
 // fluidshape type
-extern const mjMap fluid_map[2] = {
+const mjMap fluid_map[2] = {
   {"none",      0},
   {"ellipsoid", 1}
 };
 
 
 // enable type
-extern const mjMap enable_map[2] = {
+const mjMap enable_map[2] = {
   {"disable", 0},
   {"enable",  1}
 };
 
 
 // TFAuto type
-extern const mjMap TFAuto_map[3] = {
+const mjMap TFAuto_map[3] = {
   {"false",   0},
   {"true",    1},
   {"auto",    2}
@@ -652,8 +652,8 @@ const mjMap FAuto_map[2] = {
 
 
 // body sleep type
-extern const int bodysleep_sz = 4;
-extern const mjMap bodysleep_map[bodysleep_sz] = {
+const int bodysleep_sz = 4;
+const mjMap bodysleep_map[bodysleep_sz] = {
   {"auto",          mjSLEEP_AUTO},
   {"never",         mjSLEEP_NEVER},
   {"allowed",       mjSLEEP_ALLOWED},
@@ -661,8 +661,8 @@ extern const mjMap bodysleep_map[bodysleep_sz] = {
 };
 
 // joint type
-extern const int joint_sz = 4;
-extern const mjMap joint_map[joint_sz] = {
+const int joint_sz = 4;
+const mjMap joint_map[joint_sz] = {
   {"free",          mjJNT_FREE},
   {"ball",          mjJNT_BALL},
   {"slide",         mjJNT_SLIDE},
@@ -671,7 +671,7 @@ extern const mjMap joint_map[joint_sz] = {
 
 
 // geom type
-extern const mjMap geom_map[mjNGEOMTYPES] = {
+const mjMap geom_map[mjNGEOMTYPES] = {
   {"plane",         mjGEOM_PLANE},
   {"hfield",        mjGEOM_HFIELD},
   {"sphere",        mjGEOM_SPHERE},
@@ -685,15 +685,15 @@ extern const mjMap geom_map[mjNGEOMTYPES] = {
 
 
 // projection type
-extern const int projection_sz = 2;
-extern const mjMap projection_map[projection_sz] = {
+const int projection_sz = 2;
+const mjMap projection_map[projection_sz] = {
   {"perspective",   mjPROJ_PERSPECTIVE},
   {"orthographic",  mjPROJ_ORTHOGRAPHIC}
 };
 
 // camlight type
-extern const int camlight_sz = 5;
-extern const mjMap camlight_map[camlight_sz] = {
+const int camlight_sz = 5;
+const mjMap camlight_map[camlight_sz] = {
   {"fixed",         mjCAMLIGHT_FIXED},
   {"track",         mjCAMLIGHT_TRACK},
   {"trackcom",      mjCAMLIGHT_TRACKCOM},
@@ -703,8 +703,8 @@ extern const mjMap camlight_map[camlight_sz] = {
 
 
 // light type
-extern const int lighttype_sz = 4;
-extern const mjMap lighttype_map[lighttype_sz] = {
+const int lighttype_sz = 4;
+const mjMap lighttype_map[lighttype_sz] = {
   {"spot",          mjLIGHT_SPOT},
   {"directional",   mjLIGHT_DIRECTIONAL},
   {"point",         mjLIGHT_POINT},
@@ -713,8 +713,8 @@ extern const mjMap lighttype_map[lighttype_sz] = {
 
 
 // texmat role type
-extern const int texrole_sz = mjNTEXROLE - 1;
-extern const mjMap texrole_map[texrole_sz] = {
+const int texrole_sz = mjNTEXROLE - 1;
+const mjMap texrole_map[texrole_sz] = {
   {"rgb",           mjTEXROLE_RGB},
   {"occlusion",     mjTEXROLE_OCCLUSION},
   {"roughness",     mjTEXROLE_ROUGHNESS},
@@ -728,8 +728,8 @@ extern const mjMap texrole_map[texrole_sz] = {
 
 
 // integrator type
-extern const int integrator_sz = 4;
-extern const mjMap integrator_map[integrator_sz] = {
+const int integrator_sz = 4;
+const mjMap integrator_map[integrator_sz] = {
   {"Euler",         mjINT_EULER},
   {"RK4",           mjINT_RK4},
   {"implicit",      mjINT_IMPLICIT},
@@ -738,16 +738,16 @@ extern const mjMap integrator_map[integrator_sz] = {
 
 
 // cone type
-extern const int cone_sz = 2;
-extern const mjMap cone_map[cone_sz] = {
+const int cone_sz = 2;
+const mjMap cone_map[cone_sz] = {
   {"pyramidal",     mjCONE_PYRAMIDAL},
   {"elliptic",      mjCONE_ELLIPTIC}
 };
 
 
 // Jacobian type
-extern const int jac_sz = 3;
-extern const mjMap jac_map[jac_sz] = {
+const int jac_sz = 3;
+const mjMap jac_map[jac_sz] = {
   {"dense",         mjJAC_DENSE},
   {"sparse",        mjJAC_SPARSE},
   {"auto",          mjJAC_AUTO}
@@ -755,8 +755,8 @@ extern const mjMap jac_map[jac_sz] = {
 
 
 // solver type
-extern const int solver_sz = 3;
-extern const mjMap solver_map[solver_sz] = {
+const int solver_sz = 3;
+const mjMap solver_map[solver_sz] = {
   {"PGS",           mjSOL_PGS},
   {"CG",            mjSOL_CG},
   {"Newton",        mjSOL_NEWTON}
@@ -764,8 +764,8 @@ extern const mjMap solver_map[solver_sz] = {
 
 
 // constraint type
-extern const int equality_sz = 8;
-extern const mjMap equality_map[equality_sz] = {
+const int equality_sz = 8;
+const mjMap equality_map[equality_sz] = {
   {"connect",       mjEQ_CONNECT},
   {"weld",          mjEQ_WELD},
   {"joint",         mjEQ_JOINT},
@@ -778,8 +778,8 @@ extern const mjMap equality_map[equality_sz] = {
 
 
 // type for texture
-extern const int texture_sz = 3;
-extern const mjMap texture_map[texture_sz] = {
+const int texture_sz = 3;
+const mjMap texture_map[texture_sz] = {
   {"2d",            mjTEXTURE_2D},
   {"cube",          mjTEXTURE_CUBE},
   {"skybox",        mjTEXTURE_SKYBOX}
@@ -787,8 +787,8 @@ extern const mjMap texture_map[texture_sz] = {
 
 
 // colorspace for texture
-extern const int colorspace_sz = 3;
-extern const mjMap colorspace_map[colorspace_sz] = {
+const int colorspace_sz = 3;
+const mjMap colorspace_map[colorspace_sz] = {
   {"auto",          mjCOLORSPACE_AUTO},
   {"linear",        mjCOLORSPACE_LINEAR},
   {"sRGB",          mjCOLORSPACE_SRGB}
@@ -796,8 +796,8 @@ extern const mjMap colorspace_map[colorspace_sz] = {
 
 
 // builtin type for texture
-extern const int builtin_sz = 4;
-extern const mjMap builtin_map[builtin_sz] = {
+const int builtin_sz = 4;
+const mjMap builtin_map[builtin_sz] = {
   {"none",          mjBUILTIN_NONE},
   {"gradient",      mjBUILTIN_GRADIENT},
   {"checker",       mjBUILTIN_CHECKER},
@@ -806,8 +806,8 @@ extern const mjMap builtin_map[builtin_sz] = {
 
 
 // mark type for texture
-extern const int mark_sz = 4;
-extern const mjMap mark_map[mark_sz] = {
+const int mark_sz = 4;
+const mjMap mark_map[mark_sz] = {
   {"none",          mjMARK_NONE},
   {"edge",          mjMARK_EDGE},
   {"cross",         mjMARK_CROSS},
@@ -816,8 +816,8 @@ extern const mjMap mark_map[mark_sz] = {
 
 
 // dyn type
-extern const int dyn_sz = 7;
-extern const mjMap dyn_map[dyn_sz] = {
+const int dyn_sz = 7;
+const mjMap dyn_map[dyn_sz] = {
   {"none",          mjDYN_NONE},
   {"integrator",    mjDYN_INTEGRATOR},
   {"filter",        mjDYN_FILTER},
@@ -829,8 +829,8 @@ extern const mjMap dyn_map[dyn_sz] = {
 
 
 // dcmotor controller input mode
-extern const int dcmotorinput_sz = 3;
-extern const mjMap dcmotorinput_map[dcmotorinput_sz] = {
+const int dcmotorinput_sz = 3;
+const mjMap dcmotorinput_map[dcmotorinput_sz] = {
   {"voltage",       0},
   {"position",      1},
   {"velocity",      2}
@@ -838,8 +838,8 @@ extern const mjMap dcmotorinput_map[dcmotorinput_sz] = {
 
 
 // gain type
-extern const int gain_sz = 5;
-extern const mjMap gain_map[gain_sz] = {
+const int gain_sz = 5;
+const mjMap gain_map[gain_sz] = {
   {"fixed",         mjGAIN_FIXED},
   {"affine",        mjGAIN_AFFINE},
   {"muscle",        mjGAIN_MUSCLE},
@@ -849,8 +849,8 @@ extern const mjMap gain_map[gain_sz] = {
 
 
 // bias type
-extern const int bias_sz = 5;
-extern const mjMap bias_map[bias_sz] = {
+const int bias_sz = 5;
+const mjMap bias_map[bias_sz] = {
   {"none",          mjBIAS_NONE},
   {"affine",        mjBIAS_AFFINE},
   {"muscle",        mjBIAS_MUSCLE},
@@ -860,8 +860,8 @@ extern const mjMap bias_map[bias_sz] = {
 
 
 // interpolation type
-extern const int interp_sz = 3;
-extern const mjMap interp_map[interp_sz] = {
+const int interp_sz = 3;
+const mjMap interp_map[interp_sz] = {
   {"zoh",           0},
   {"linear",        1},
   {"cubic",         2}
@@ -869,8 +869,8 @@ extern const mjMap interp_map[interp_sz] = {
 
 
 // stage type
-extern const int stage_sz = 4;
-extern const mjMap stage_map[stage_sz] = {
+const int stage_sz = 4;
+const mjMap stage_map[stage_sz] = {
   {"none",          mjSTAGE_NONE},
   {"pos",           mjSTAGE_POS},
   {"vel",           mjSTAGE_VEL},
@@ -879,8 +879,8 @@ extern const mjMap stage_map[stage_sz] = {
 
 
 // data type
-extern const int datatype_sz = 4;
-extern const mjMap datatype_map[datatype_sz] = {
+const int datatype_sz = 4;
+const mjMap datatype_map[datatype_sz] = {
   {"real",          mjDATATYPE_REAL},
   {"positive",      mjDATATYPE_POSITIVE},
   {"axis",          mjDATATYPE_AXIS},
@@ -889,7 +889,7 @@ extern const mjMap datatype_map[datatype_sz] = {
 
 
 // contact data type
-extern const mjMap condata_map[mjNCONDATA] = {
+const mjMap condata_map[mjNCONDATA] = {
   {"found",         mjCONDATA_FOUND},
   {"force",         mjCONDATA_FORCE},
   {"torque",        mjCONDATA_TORQUE},
@@ -901,7 +901,7 @@ extern const mjMap condata_map[mjNCONDATA] = {
 
 
 // rangefinder data type
-extern const mjMap raydata_map[mjNRAYDATA] = {
+const mjMap raydata_map[mjNRAYDATA] = {
   {"dist",          mjRAYDATA_DIST},
   {"dir",           mjRAYDATA_DIR},
   {"origin",        mjRAYDATA_ORIGIN},
@@ -911,16 +911,16 @@ extern const mjMap raydata_map[mjNRAYDATA] = {
 };
 
 // camera output type
-extern const int camout_sz = mjNCAMOUT;
-extern const mjMap camout_map[mjNCAMOUT] = {{"rgb", mjCAMOUT_RGB},
+const int camout_sz = mjNCAMOUT;
+const mjMap camout_map[mjNCAMOUT] = {{"rgb", mjCAMOUT_RGB},
                                      {"depth", mjCAMOUT_DEPTH},
                                      {"distance", mjCAMOUT_DIST},
                                      {"normal", mjCAMOUT_NORMAL},
                                      {"segmentation", mjCAMOUT_SEG}};
 
 // contact reduction type
-extern const int reduce_sz = 4;
-extern const mjMap reduce_map[reduce_sz] = {
+const int reduce_sz = 4;
+const mjMap reduce_map[reduce_sz] = {
   {"none",          0},
   {"mindist",       1},
   {"maxforce",      2},
@@ -934,8 +934,8 @@ const mjMap conflict_map[conflict_sz] = {{"warning", mjCONFLICT_WARNING},
                                          {"error", mjCONFLICT_ERROR}};
 
 // LR mode
-extern const int lrmode_sz = 4;
-extern const mjMap lrmode_map[lrmode_sz] = {
+const int lrmode_sz = 4;
+const mjMap lrmode_map[lrmode_sz] = {
   {"none",          mjLRMODE_NONE},
   {"muscle",        mjLRMODE_MUSCLE},
   {"muscleuser",    mjLRMODE_MUSCLEUSER},
@@ -944,7 +944,7 @@ extern const mjMap lrmode_map[lrmode_sz] = {
 
 
 // composite type
-extern const mjMap comp_map[mjNCOMPTYPES] = {
+const mjMap comp_map[mjNCOMPTYPES] = {
   {"particle",      mjCOMPTYPE_PARTICLE},
   {"grid",          mjCOMPTYPE_GRID},
   {"rope",          mjCOMPTYPE_ROPE},
@@ -955,13 +955,13 @@ extern const mjMap comp_map[mjNCOMPTYPES] = {
 
 
 // composite joint kind
-extern const mjMap jkind_map[1] = {
+const mjMap jkind_map[1] = {
   {"main",          mjCOMPKIND_JOINT}
 };
 
 
 // composite rope shape
-extern const mjMap shape_map[mjNCOMPSHAPES] = {
+const mjMap shape_map[mjNCOMPSHAPES] = {
   {"s",             mjCOMPSHAPE_LINE},
   {"cos(s)",        mjCOMPSHAPE_COS},
   {"sin(s)",        mjCOMPSHAPE_SIN},
@@ -970,14 +970,14 @@ extern const mjMap shape_map[mjNCOMPSHAPES] = {
 
 
 // mesh type
-extern const mjMap meshtype_map[2] = {
+const mjMap meshtype_map[2] = {
   {"false",         mjINERTIA_VOLUME},
   {"true",          mjINERTIA_SHELL},
 };
 
 
 // mesh inertia type
-extern const mjMap meshinertia_map[4] = {
+const mjMap meshinertia_map[4] = {
   {"convex",        mjMESH_INERTIA_CONVEX},
   {"legacy",        mjMESH_INERTIA_LEGACY},
   {"exact",         mjMESH_INERTIA_EXACT},
@@ -986,8 +986,8 @@ extern const mjMap meshinertia_map[4] = {
 
 
 // mesh builtin type
-extern const int meshbuiltin_sz = 8;
-extern const mjMap meshbuiltin_map[meshbuiltin_sz] = {
+const int meshbuiltin_sz = 8;
+const mjMap meshbuiltin_map[meshbuiltin_sz] = {
   {"none",          mjMESH_BUILTIN_NONE},
   {"sphere",        mjMESH_BUILTIN_SPHERE},
   {"hemisphere",    mjMESH_BUILTIN_HEMISPHERE},
@@ -1000,7 +1000,7 @@ extern const mjMap meshbuiltin_map[meshbuiltin_sz] = {
 
 
 // flexcomp type
-extern const mjMap fcomp_map[mjNFCOMPTYPES] = {
+const mjMap fcomp_map[mjNFCOMPTYPES] = {
   {"grid",          mjFCOMPTYPE_GRID},
   {"box",           mjFCOMPTYPE_BOX},
   {"cylinder",      mjFCOMPTYPE_CYLINDER},
@@ -1015,7 +1015,7 @@ extern const mjMap fcomp_map[mjNFCOMPTYPES] = {
 
 
 // flexcomp dof type
-extern const mjMap fdof_map[mjNFCOMPDOFS] = {
+const mjMap fdof_map[mjNFCOMPDOFS] = {
   {"full",          mjFCOMPDOF_FULL},
   {"radial",        mjFCOMPDOF_RADIAL},
   {"trilinear",     mjFCOMPDOF_TRILINEAR},
@@ -1025,7 +1025,7 @@ extern const mjMap fdof_map[mjNFCOMPDOFS] = {
 
 
 // flex selfcollide type
-extern const mjMap flexself_map[5] = {
+const mjMap flexself_map[5] = {
   {"none",          mjFLEXSELF_NONE},
   {"narrow",        mjFLEXSELF_NARROW},
   {"bvh",           mjFLEXSELF_BVH},
@@ -1035,7 +1035,7 @@ extern const mjMap flexself_map[5] = {
 
 
 // flex elastic 2d type
-extern const mjMap elastic2d_map[5] = {
+const mjMap elastic2d_map[5] = {
   {"none",          0},
   {"bend",          1},
   {"stretch",       2},
@@ -1044,13 +1044,1041 @@ extern const mjMap elastic2d_map[5] = {
 
 
 // flex equality type
-extern const mjMap flexeq_map[4] = {
+const mjMap flexeq_map[4] = {
   {"false",         0},
   {"true",          1},
   {"vert",          2},
   {"strain",        3},
 };
 
+
+//---------------------------------- Attribute type table ------------------------------------------
+// Typed mirror of the ReadAttr*/MapValue calls in the OneX() parsers below.
+// Consumed by the XSD generator in xml_native_schema.cc.
+//
+// MAINTENANCE: if you add/rename an attribute in OneX() or MJCF[], add the
+// matching row here too (one per (element, attr) pair).
+
+// clang-format off
+const mjXAttr kMjXAttrTable[] = {
+
+  //-------------------------------- <mujoco> ------------------------------------------------------
+  {"mujoco", "model", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <compiler> ----------------------------------------------------
+  {"compiler", "autolimits",        kMjXEnum,   0, bool_map,        2},
+  {"compiler", "boundmass",         kMjXReal,   0, nullptr,         0},
+  {"compiler", "boundinertia",      kMjXReal,   0, nullptr,         0},
+  {"compiler", "settotalmass",      kMjXReal,   0, nullptr,         0},
+  {"compiler", "balanceinertia",    kMjXEnum,   0, bool_map,        2},
+  {"compiler", "strippath",         kMjXEnum,   0, bool_map,        2},
+  {"compiler", "coordinate",        kMjXEnum,   0, coordinate_map,  2},
+  {"compiler", "angle",             kMjXEnum,   0, angle_map,       2},
+  {"compiler", "fitaabb",           kMjXEnum,   0, bool_map,        2},
+  {"compiler", "eulerseq",          kMjXString, 0, nullptr,         0},
+  {"compiler", "meshdir",           kMjXString, 0, nullptr,         0},
+  {"compiler", "texturedir",        kMjXString, 0, nullptr,         0},
+  {"compiler", "discardvisual",     kMjXEnum,   0, bool_map,        2},
+  {"compiler", "usethread",         kMjXEnum,   0, bool_map,        2},
+  {"compiler", "fusestatic",        kMjXEnum,   0, bool_map,        2},
+  {"compiler", "inertiafromgeom",   kMjXEnum,   0, TFAuto_map,      3},
+  {"compiler", "inertiagrouprange", kMjXIntN,   2, nullptr,         0},
+  {"compiler", "saveinertial",      kMjXEnum,   0, bool_map,        2},
+  {"compiler", "assetdir",          kMjXString, 0, nullptr,         0},
+  {"compiler", "alignfree",         kMjXEnum,   0, bool_map,        2},
+  //-------------------------------- <compiler><lengthrange> ---------------------------------------
+  {"lengthrange", "mode",        kMjXEnum, 0, lrmode_map, 4},
+  {"lengthrange", "useexisting", kMjXEnum, 0, bool_map,   2},
+  {"lengthrange", "uselimit",    kMjXEnum, 0, bool_map,   2},
+  {"lengthrange", "accel",       kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "maxforce",    kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "timeconst",   kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "timestep",    kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "inttotal",    kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "interval",    kMjXReal, 0, nullptr,    0},
+  {"lengthrange", "tolrange",    kMjXReal, 0, nullptr,    0},
+
+  //-------------------------------- <option> ------------------------------------------------------
+  {"option", "timestep",             kMjXReal,      0, nullptr,        0},
+  {"option", "impratio",             kMjXReal,      0, nullptr,        0},
+  {"option", "tolerance",            kMjXReal,      0, nullptr,        0},
+  {"option", "ls_tolerance",         kMjXReal,      0, nullptr,        0},
+  {"option", "noslip_tolerance",     kMjXReal,      0, nullptr,        0},
+  {"option", "ccd_tolerance",        kMjXReal,      0, nullptr,        0},
+  {"option", "sleep_tolerance",      kMjXReal,      0, nullptr,        0},
+  {"option", "gravity",              kMjXRealN,     3, nullptr,        0},
+  {"option", "wind",                 kMjXRealN,     3, nullptr,        0},
+  {"option", "magnetic",             kMjXRealN,     3, nullptr,        0},
+  {"option", "density",              kMjXReal,      0, nullptr,        0},
+  {"option", "viscosity",            kMjXReal,      0, nullptr,        0},
+  {"option", "o_margin",             kMjXReal,      0, nullptr,        0},
+  {"option", "o_solref",             kMjXRealUpToN, 2, nullptr,        0},
+  {"option", "o_solimp",             kMjXRealUpToN, 5, nullptr,        0},
+  {"option", "o_friction",           kMjXRealUpToN, 5, nullptr,        0},
+  {"option", "integrator",           kMjXEnum,      0, integrator_map, 4},
+  {"option", "cone",                 kMjXEnum,      0, cone_map,       2},
+  {"option", "jacobian",             kMjXEnum,      0, jac_map,        3},
+  {"option", "solver",               kMjXEnum,      0, solver_map,     3},
+  {"option", "iterations",           kMjXInt,       0, nullptr,        0},
+  {"option", "ls_iterations",        kMjXInt,       0, nullptr,        0},
+  {"option", "noslip_iterations",    kMjXInt,       0, nullptr,        0},
+  {"option", "ccd_iterations",       kMjXInt,       0, nullptr,        0},
+  {"option", "sdf_iterations",       kMjXInt,       0, nullptr,        0},
+  {"option", "sdf_initpoints",       kMjXInt,       0, nullptr,        0},
+  {"option", "actuatorgroupdisable", kMjXIntVec,    0, nullptr,        0},
+
+  //-------------------------------- <option><flag> ------------------------------------------------
+  {"flag", "constraint",   kMjXEnum, 0, enable_map, 2},
+  {"flag", "equality",     kMjXEnum, 0, enable_map, 2},
+  {"flag", "frictionloss", kMjXEnum, 0, enable_map, 2},
+  {"flag", "limit",        kMjXEnum, 0, enable_map, 2},
+  {"flag", "contact",      kMjXEnum, 0, enable_map, 2},
+  {"flag", "spring",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "damper",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "gravity",      kMjXEnum, 0, enable_map, 2},
+  {"flag", "clampctrl",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "warmstart",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "filterparent", kMjXEnum, 0, enable_map, 2},
+  {"flag", "actuation",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "refsafe",      kMjXEnum, 0, enable_map, 2},
+  {"flag", "sensor",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "midphase",     kMjXEnum, 0, enable_map, 2},
+  {"flag", "eulerdamp",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "autoreset",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "nativeccd",    kMjXEnum, 0, enable_map, 2},
+  {"flag", "island",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "override",     kMjXEnum, 0, enable_map, 2},
+  {"flag", "energy",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "fwdinv",       kMjXEnum, 0, enable_map, 2},
+  {"flag", "invdiscrete",  kMjXEnum, 0, enable_map, 2},
+  {"flag", "multiccd",     kMjXEnum, 0, enable_map, 2},
+  {"flag", "sleep",        kMjXEnum, 0, enable_map, 2},
+
+  //-------------------------------- <size> --------------------------------------------------------
+  {"size", "memory",         kMjXString, 0, nullptr, 0},
+  {"size", "njmax",          kMjXInt,    0, nullptr, 0},
+  {"size", "nconmax",        kMjXInt,    0, nullptr, 0},
+  {"size", "nstack",         kMjXInt,    0, nullptr, 0},
+  {"size", "nuserdata",      kMjXInt,    0, nullptr, 0},
+  {"size", "nkey",           kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_body",     kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_jnt",      kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_geom",     kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_site",     kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_cam",      kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_tendon",   kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_actuator", kMjXInt,    0, nullptr, 0},
+  {"size", "nuser_sensor",   kMjXInt,    0, nullptr, 0},
+
+  //-------------------------------- <statistic> ---------------------------------------------------
+  {"statistic", "meaninertia", kMjXReal,  0, nullptr, 0},
+  {"statistic", "meanmass",    kMjXReal,  0, nullptr, 0},
+  {"statistic", "meansize",    kMjXReal,  0, nullptr, 0},
+  {"statistic", "extent",      kMjXReal,  0, nullptr, 0},
+  {"statistic", "center",      kMjXRealN, 3, nullptr, 0},
+
+  //-------------------------------- <visual> subelements ------------------------------------------
+  {"global", "cameraid",         kMjXInt,  0, nullptr,  0},
+  {"global", "orthographic",     kMjXEnum, 0, bool_map, 2},
+  {"global", "fovy",             kMjXReal, 0, nullptr,  0},
+  {"global", "ipd",              kMjXReal, 0, nullptr,  0},
+  {"global", "azimuth",          kMjXReal, 0, nullptr,  0},
+  {"global", "elevation",        kMjXReal, 0, nullptr,  0},
+  {"global", "linewidth",        kMjXReal, 0, nullptr,  0},
+  {"global", "glow",             kMjXReal, 0, nullptr,  0},
+  {"global", "offwidth",         kMjXInt,  0, nullptr,  0},
+  {"global", "offheight",        kMjXInt,  0, nullptr,  0},
+  {"global", "realtime",         kMjXReal, 0, nullptr,  0},
+  {"global", "ellipsoidinertia", kMjXEnum, 0, bool_map, 2},
+  {"global", "bvactive",         kMjXEnum, 0, bool_map, 2},
+
+  {"quality", "shadowsize", kMjXInt, 0, nullptr, 0},
+  {"quality", "offsamples", kMjXInt, 0, nullptr, 0},
+  {"quality", "numslices",  kMjXInt, 0, nullptr, 0},
+  {"quality", "numstacks",  kMjXInt, 0, nullptr, 0},
+  {"quality", "numquads",   kMjXInt, 0, nullptr, 0},
+
+  {"headlight", "ambient",  kMjXRealN, 3, nullptr, 0},
+  {"headlight", "diffuse",  kMjXRealN, 3, nullptr, 0},
+  {"headlight", "specular", kMjXRealN, 3, nullptr, 0},
+  {"headlight", "active",   kMjXInt,   0, nullptr, 0},
+
+  {"map", "stiffness",      kMjXReal, 0, nullptr, 0},
+  {"map", "stiffnessrot",   kMjXReal, 0, nullptr, 0},
+  {"map", "force",          kMjXReal, 0, nullptr, 0},
+  {"map", "torque",         kMjXReal, 0, nullptr, 0},
+  {"map", "alpha",          kMjXReal, 0, nullptr, 0},
+  {"map", "fogstart",       kMjXReal, 0, nullptr, 0},
+  {"map", "fogend",         kMjXReal, 0, nullptr, 0},
+  {"map", "znear",          kMjXReal, 0, nullptr, 0},
+  {"map", "zfar",           kMjXReal, 0, nullptr, 0},
+  {"map", "haze",           kMjXReal, 0, nullptr, 0},
+  {"map", "shadowclip",     kMjXReal, 0, nullptr, 0},
+  {"map", "shadowscale",    kMjXReal, 0, nullptr, 0},
+  {"map", "actuatortendon", kMjXReal, 0, nullptr, 0},
+
+  {"scale", "forcewidth",     kMjXReal, 0, nullptr, 0},
+  {"scale", "contactwidth",   kMjXReal, 0, nullptr, 0},
+  {"scale", "contactheight",  kMjXReal, 0, nullptr, 0},
+  {"scale", "connect",        kMjXReal, 0, nullptr, 0},
+  {"scale", "com",            kMjXReal, 0, nullptr, 0},
+  {"scale", "camera",         kMjXReal, 0, nullptr, 0},
+  {"scale", "light",          kMjXReal, 0, nullptr, 0},
+  {"scale", "selectpoint",    kMjXReal, 0, nullptr, 0},
+  {"scale", "jointlength",    kMjXReal, 0, nullptr, 0},
+  {"scale", "jointwidth",     kMjXReal, 0, nullptr, 0},
+  {"scale", "actuatorlength", kMjXReal, 0, nullptr, 0},
+  {"scale", "actuatorwidth",  kMjXReal, 0, nullptr, 0},
+  {"scale", "framelength",    kMjXReal, 0, nullptr, 0},
+  {"scale", "framewidth",     kMjXReal, 0, nullptr, 0},
+  {"scale", "constraint",     kMjXReal, 0, nullptr, 0},
+  {"scale", "slidercrank",    kMjXReal, 0, nullptr, 0},
+  {"scale", "frustum",        kMjXReal, 0, nullptr, 0},
+
+  {"rgba", "fog",              kMjXRealN, 4, nullptr, 0},
+  {"rgba", "haze",             kMjXRealN, 4, nullptr, 0},
+  {"rgba", "force",            kMjXRealN, 4, nullptr, 0},
+  {"rgba", "inertia",          kMjXRealN, 4, nullptr, 0},
+  {"rgba", "joint",            kMjXRealN, 4, nullptr, 0},
+  {"rgba", "actuator",         kMjXRealN, 4, nullptr, 0},
+  {"rgba", "actuatornegative", kMjXRealN, 4, nullptr, 0},
+  {"rgba", "actuatorpositive", kMjXRealN, 4, nullptr, 0},
+  {"rgba", "com",              kMjXRealN, 4, nullptr, 0},
+  {"rgba", "camera",           kMjXRealN, 4, nullptr, 0},
+  {"rgba", "light",            kMjXRealN, 4, nullptr, 0},
+  {"rgba", "selectpoint",      kMjXRealN, 4, nullptr, 0},
+  {"rgba", "connect",          kMjXRealN, 4, nullptr, 0},
+  {"rgba", "contactpoint",     kMjXRealN, 4, nullptr, 0},
+  {"rgba", "contactforce",     kMjXRealN, 4, nullptr, 0},
+  {"rgba", "contactfriction",  kMjXRealN, 4, nullptr, 0},
+  {"rgba", "contacttorque",    kMjXRealN, 4, nullptr, 0},
+  {"rgba", "contactgap",       kMjXRealN, 4, nullptr, 0},
+  {"rgba", "rangefinder",      kMjXRealN, 4, nullptr, 0},
+  {"rgba", "constraint",       kMjXRealN, 4, nullptr, 0},
+  {"rgba", "slidercrank",      kMjXRealN, 4, nullptr, 0},
+  {"rgba", "crankbroken",      kMjXRealN, 4, nullptr, 0},
+  {"rgba", "frustum",          kMjXRealN, 4, nullptr, 0},
+  {"rgba", "bv",               kMjXRealN, 4, nullptr, 0},
+  {"rgba", "bvactive",         kMjXRealN, 4, nullptr, 0},
+
+  //-------------------------------- <default> -----------------------------------------------------
+  {"default", "class", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <mesh> (asset + default) -------------------------------------
+  {"mesh", "name",         kMjXString,    0, nullptr,          0},
+  {"mesh", "class",        kMjXString,    0, nullptr,          0},
+  {"mesh", "content_type", kMjXString,    0, nullptr,          0},
+  {"mesh", "file",         kMjXString,    0, nullptr,          0},
+  {"mesh", "vertex",       kMjXRealVec,   0, nullptr,          0},
+  {"mesh", "normal",       kMjXRealVec,   0, nullptr,          0},
+  {"mesh", "texcoord",     kMjXRealVec,   0, nullptr,          0},
+  {"mesh", "face",         kMjXIntVec,    0, nullptr,          0},
+  {"mesh", "refpos",       kMjXRealN,     3, nullptr,          0},
+  {"mesh", "refquat",      kMjXRealN,     4, nullptr,          0},
+  {"mesh", "scale",        kMjXRealN,     3, nullptr,          0},
+  {"mesh", "smoothnormal", kMjXEnum,      0, bool_map,         2},
+  {"mesh", "maxhullvert",  kMjXInt,       0, nullptr,          0},
+  {"mesh", "inertia",      kMjXEnum,      0, meshinertia_map,  4},
+  {"mesh", "builtin",      kMjXEnum,      0, meshbuiltin_map,  8},
+  {"mesh", "params",       kMjXRealVec,   0, nullptr,          0},
+  {"mesh", "material",     kMjXString,    0, nullptr,          0},
+
+  //-------------------------------- <material> (asset + default) ----------------------------------
+  {"material", "name",        kMjXString, 0, nullptr,  0},
+  {"material", "class",       kMjXString, 0, nullptr,  0},
+  {"material", "texture",     kMjXString, 0, nullptr,  0},
+  {"material", "texrepeat",   kMjXRealN,  2, nullptr,  0},
+  {"material", "texuniform",  kMjXEnum,   0, bool_map, 2},
+  {"material", "emission",    kMjXReal,   0, nullptr,  0},
+  {"material", "specular",    kMjXReal,   0, nullptr,  0},
+  {"material", "shininess",   kMjXReal,   0, nullptr,  0},
+  {"material", "reflectance", kMjXReal,   0, nullptr,  0},
+  {"material", "metallic",    kMjXReal,   0, nullptr,  0},
+  {"material", "roughness",   kMjXReal,   0, nullptr,  0},
+  {"material", "rgba",        kMjXRealN,  4, nullptr,  0},
+
+  //-------------------------------- <material><layer> ---------------------------------------------
+  {"layer", "texture", kMjXString, 0, nullptr,     0},
+  {"layer", "role",    kMjXEnum,   0, texrole_map, 9},
+
+  //-------------------------------- <joint> (body + default) --------------------------------------
+  {"joint", "name",              kMjXString,    0, nullptr,    0},
+  {"joint", "class",             kMjXString,    0, nullptr,    0},
+  {"joint", "type",              kMjXEnum,      0, joint_map,  4},
+  {"joint", "group",             kMjXInt,       0, nullptr,    0},
+  {"joint", "pos",               kMjXRealN,     3, nullptr,    0},
+  {"joint", "axis",              kMjXRealN,     3, nullptr,    0},
+  {"joint", "springdamper",      kMjXRealN,     2, nullptr,    0},
+  {"joint", "limited",           kMjXEnum,      0, TFAuto_map, 3},
+  {"joint", "actuatorfrclimited",kMjXEnum,      0, TFAuto_map, 3},
+  {"joint", "solreflimit",       kMjXRealUpToN, 2, nullptr,    0},
+  {"joint", "solimplimit",       kMjXRealUpToN, 5, nullptr,    0},
+  {"joint", "solreffriction",    kMjXRealUpToN, 2, nullptr,    0},
+  {"joint", "solimpfriction",    kMjXRealUpToN, 5, nullptr,    0},
+  {"joint", "stiffness",         kMjXRealUpToN, 3, nullptr,    0},
+  {"joint", "range",             kMjXRealN,     2, nullptr,    0},
+  {"joint", "actuatorfrcrange",  kMjXRealN,     2, nullptr,    0},
+  {"joint", "actuatorgravcomp",  kMjXEnum,      0, bool_map,   2},
+  {"joint", "margin",            kMjXReal,      0, nullptr,    0},
+  {"joint", "ref",               kMjXReal,      0, nullptr,    0},
+  {"joint", "springref",         kMjXReal,      0, nullptr,    0},
+  {"joint", "armature",          kMjXReal,      0, nullptr,    0},
+  {"joint", "damping",           kMjXRealUpToN, 3, nullptr,    0},
+  {"joint", "frictionloss",      kMjXReal,      0, nullptr,    0},
+  {"joint", "user",              kMjXRealVec,   0, nullptr,    0},
+
+  //-------------------------------- <freejoint> (body) --------------------------------------------
+  {"freejoint", "name",  kMjXString, 0, nullptr,    0},
+  {"freejoint", "group", kMjXInt,    0, nullptr,    0},
+  {"freejoint", "align", kMjXEnum,   0, TFAuto_map, 3},
+
+  //-------------------------------- <geom> (body + default) ---------------------------------------
+  {"geom", "name",         kMjXString,    0, nullptr,   0},
+  {"geom", "class",        kMjXString,    0, nullptr,   0},
+  {"geom", "type",         kMjXEnum,      0, geom_map,  mjNGEOMTYPES},
+  {"geom", "pos",          kMjXRealN,     3, nullptr,   0},
+  {"geom", "quat",         kMjXRealN,     4, nullptr,   0},
+  {"geom", "axisangle",    kMjXRealN,     4, nullptr,   0},
+  {"geom", "xyaxes",       kMjXRealN,     6, nullptr,   0},
+  {"geom", "zaxis",        kMjXRealN,     3, nullptr,   0},
+  {"geom", "euler",        kMjXRealN,     3, nullptr,   0},
+  {"geom", "contype",      kMjXInt,       0, nullptr,   0},
+  {"geom", "conaffinity",  kMjXInt,       0, nullptr,   0},
+  {"geom", "condim",       kMjXInt,       0, nullptr,   0},
+  {"geom", "group",        kMjXInt,       0, nullptr,   0},
+  {"geom", "priority",     kMjXInt,       0, nullptr,   0},
+  {"geom", "size",         kMjXRealUpToN, 3, nullptr,   0},
+  {"geom", "material",     kMjXString,    0, nullptr,   0},
+  {"geom", "friction",     kMjXRealUpToN, 3, nullptr,   0},
+  {"geom", "mass",         kMjXReal,      0, nullptr,   0},
+  {"geom", "density",      kMjXReal,      0, nullptr,   0},
+  {"geom", "shellinertia", kMjXEnum,      0, meshtype_map, 2},
+  {"geom", "solmix",       kMjXReal,      0, nullptr,   0},
+  {"geom", "solref",       kMjXRealUpToN, 2, nullptr,   0},
+  {"geom", "solimp",       kMjXRealUpToN, 5, nullptr,   0},
+  {"geom", "margin",       kMjXReal,      0, nullptr,   0},
+  {"geom", "gap",          kMjXReal,      0, nullptr,   0},
+  {"geom", "fromto",       kMjXRealN,     6, nullptr,   0},
+  {"geom", "hfield",       kMjXString,    0, nullptr,   0},
+  {"geom", "mesh",         kMjXString,    0, nullptr,   0},
+  {"geom", "fitscale",     kMjXReal,      0, nullptr,   0},
+  {"geom", "rgba",         kMjXRealN,     4, nullptr,   0},
+  {"geom", "fluidshape",   kMjXEnum,      0, fluid_map, 2},
+  {"geom", "fluidcoef",    kMjXRealUpToN, 5, nullptr,   0},
+  {"geom", "user",         kMjXRealVec,   0, nullptr,   0},
+
+  //-------------------------------- <site> (body + default) ---------------------------------------
+  {"site", "name",      kMjXString,    0, nullptr,  0},
+  {"site", "class",     kMjXString,    0, nullptr,  0},
+  {"site", "type",      kMjXEnum,      0, geom_map, mjNGEOMTYPES},
+  {"site", "group",     kMjXInt,       0, nullptr,  0},
+  {"site", "pos",       kMjXRealN,     3, nullptr,  0},
+  {"site", "quat",      kMjXRealN,     4, nullptr,  0},
+  {"site", "material",  kMjXString,    0, nullptr,  0},
+  {"site", "size",      kMjXRealUpToN, 3, nullptr,  0},
+  {"site", "fromto",    kMjXRealN,     6, nullptr,  0},
+  {"site", "axisangle", kMjXRealN,     4, nullptr,  0},
+  {"site", "xyaxes",    kMjXRealN,     6, nullptr,  0},
+  {"site", "zaxis",     kMjXRealN,     3, nullptr,  0},
+  {"site", "euler",     kMjXRealN,     3, nullptr,  0},
+  {"site", "rgba",      kMjXRealN,     4, nullptr,  0},
+  {"site", "user",      kMjXRealVec,   0, nullptr,  0},
+
+  //-------------------------------- <camera> (body + default) -------------------------------------
+  {"camera", "name",           kMjXString,  0, nullptr,        0},
+  {"camera", "class",          kMjXString,  0, nullptr,        0},
+  {"camera", "projection",     kMjXEnum,    0, projection_map, 2},
+  {"camera", "fovy",           kMjXReal,    0, nullptr,        0},
+  {"camera", "ipd",            kMjXReal,    0, nullptr,        0},
+  {"camera", "resolution",     kMjXIntN,    2, nullptr,        0},
+  {"camera", "output",         kMjXString,  0, nullptr,        0},  // space-separated bitflags
+  {"camera", "pos",            kMjXRealN,   3, nullptr,        0},
+  {"camera", "quat",           kMjXRealN,   4, nullptr,        0},
+  {"camera", "axisangle",      kMjXRealN,   4, nullptr,        0},
+  {"camera", "xyaxes",         kMjXRealN,   6, nullptr,        0},
+  {"camera", "zaxis",          kMjXRealN,   3, nullptr,        0},
+  {"camera", "euler",          kMjXRealN,   3, nullptr,        0},
+  {"camera", "mode",           kMjXEnum,    0, camlight_map,   5},
+  {"camera", "target",         kMjXString,  0, nullptr,        0},
+  {"camera", "focal",          kMjXRealN,   2, nullptr,        0},
+  {"camera", "focalpixel",     kMjXRealN,   2, nullptr,        0},
+  {"camera", "principal",      kMjXRealN,   2, nullptr,        0},
+  {"camera", "principalpixel", kMjXRealN,   2, nullptr,        0},
+  {"camera", "sensorsize",     kMjXRealN,   2, nullptr,        0},
+  {"camera", "user",           kMjXRealVec, 0, nullptr,        0},
+
+  //-------------------------------- <light> (body + default) --------------------------------------
+  {"light", "name",        kMjXString, 0, nullptr,       0},
+  {"light", "class",       kMjXString, 0, nullptr,       0},
+  {"light", "pos",         kMjXRealN,  3, nullptr,       0},
+  {"light", "dir",         kMjXRealN,  3, nullptr,       0},
+  {"light", "bulbradius",  kMjXReal,   0, nullptr,       0},
+  {"light", "intensity",   kMjXReal,   0, nullptr,       0},
+  {"light", "range",       kMjXReal,   0, nullptr,       0},
+  {"light", "directional", kMjXEnum,   0, bool_map,      2},
+  {"light", "type",        kMjXEnum,   0, lighttype_map, 4},
+  {"light", "castshadow",  kMjXEnum,   0, bool_map,      2},
+  {"light", "active",      kMjXEnum,   0, bool_map,      2},
+  {"light", "attenuation", kMjXRealN,  3, nullptr,       0},
+  {"light", "cutoff",      kMjXReal,   0, nullptr,       0},
+  {"light", "exponent",    kMjXReal,   0, nullptr,       0},
+  {"light", "ambient",     kMjXRealN,  3, nullptr,       0},
+  {"light", "diffuse",     kMjXRealN,  3, nullptr,       0},
+  {"light", "specular",    kMjXRealN,  3, nullptr,       0},
+  {"light", "mode",        kMjXEnum,   0, camlight_map,  5},
+  {"light", "target",      kMjXString, 0, nullptr,       0},
+  {"light", "texture",     kMjXString, 0, nullptr,       0},
+
+  //-------------------------------- <pair> (default + contact) -----------------------------------
+  {"pair", "name",           kMjXString,    0, nullptr, 0},
+  {"pair", "class",          kMjXString,    0, nullptr, 0},
+  {"pair", "geom1",          kMjXString,    0, nullptr, 0},
+  {"pair", "geom2",          kMjXString,    0, nullptr, 0},
+  {"pair", "condim",         kMjXInt,       0, nullptr, 0},
+  {"pair", "friction",       kMjXRealUpToN, 5, nullptr, 0},
+  {"pair", "solref",         kMjXRealUpToN, 2, nullptr, 0},
+  {"pair", "solreffriction", kMjXRealUpToN, 2, nullptr, 0},
+  {"pair", "solimp",         kMjXRealUpToN, 5, nullptr, 0},
+  {"pair", "margin",         kMjXReal,      0, nullptr, 0},
+  {"pair", "gap",            kMjXReal,      0, nullptr, 0},
+
+  //-------------------------------- <equality> in <default> --------------------------------------
+  {"equality", "active", kMjXEnum,      0, bool_map, 2},
+  {"equality", "solref", kMjXRealUpToN, 2, nullptr,  0},
+  {"equality", "solimp", kMjXRealUpToN, 5, nullptr,  0},
+
+  //-------------------------------- <tendon> (default + top-level tendon is a complex section) ---
+  // The <tendon> row in MJCF[] under <default> holds only tendon defaults.
+  // These attrs are shared by <spatial> and <fixed> when emitted from <tendon>.
+  {"tendon", "group",              kMjXInt,       0, nullptr,    0},
+  {"tendon", "limited",            kMjXEnum,      0, TFAuto_map, 3},
+  {"tendon", "range",              kMjXRealN,     2, nullptr,    0},
+  {"tendon", "solreflimit",        kMjXRealUpToN, 2, nullptr,    0},
+  {"tendon", "solimplimit",        kMjXRealUpToN, 5, nullptr,    0},
+  {"tendon", "solreffriction",     kMjXRealUpToN, 2, nullptr,    0},
+  {"tendon", "solimpfriction",     kMjXRealUpToN, 5, nullptr,    0},
+  {"tendon", "frictionloss",       kMjXReal,      0, nullptr,    0},
+  {"tendon", "springlength",       kMjXRealUpToN, 2, nullptr,    0},
+  {"tendon", "width",              kMjXReal,      0, nullptr,    0},
+  {"tendon", "material",           kMjXString,    0, nullptr,    0},
+  {"tendon", "margin",             kMjXReal,      0, nullptr,    0},
+  {"tendon", "stiffness",          kMjXRealUpToN, 3, nullptr,    0},
+  {"tendon", "damping",            kMjXRealUpToN, 3, nullptr,    0},
+  {"tendon", "rgba",               kMjXRealN,     4, nullptr,    0},
+  {"tendon", "user",               kMjXRealVec,   0, nullptr,    0},
+
+  //-------------------------------- <spatial> (tendon section) -----------------------------------
+  {"spatial", "name",              kMjXString,    0, nullptr,    0},
+  {"spatial", "class",             kMjXString,    0, nullptr,    0},
+  {"spatial", "group",             kMjXInt,       0, nullptr,    0},
+  {"spatial", "limited",           kMjXEnum,      0, TFAuto_map, 3},
+  {"spatial", "actuatorfrclimited",kMjXEnum,      0, TFAuto_map, 3},
+  {"spatial", "range",             kMjXRealN,     2, nullptr,    0},
+  {"spatial", "actuatorfrcrange",  kMjXRealN,     2, nullptr,    0},
+  {"spatial", "solreflimit",       kMjXRealUpToN, 2, nullptr,    0},
+  {"spatial", "solimplimit",       kMjXRealUpToN, 5, nullptr,    0},
+  {"spatial", "solreffriction",    kMjXRealUpToN, 2, nullptr,    0},
+  {"spatial", "solimpfriction",    kMjXRealUpToN, 5, nullptr,    0},
+  {"spatial", "frictionloss",      kMjXReal,      0, nullptr,    0},
+  {"spatial", "springlength",      kMjXRealUpToN, 2, nullptr,    0},
+  {"spatial", "width",             kMjXReal,      0, nullptr,    0},
+  {"spatial", "material",          kMjXString,    0, nullptr,    0},
+  {"spatial", "margin",            kMjXReal,      0, nullptr,    0},
+  {"spatial", "stiffness",         kMjXRealUpToN, 3, nullptr,    0},
+  {"spatial", "damping",           kMjXRealUpToN, 3, nullptr,    0},
+  {"spatial", "armature",          kMjXReal,      0, nullptr,    0},
+  {"spatial", "rgba",              kMjXRealN,     4, nullptr,    0},
+  {"spatial", "user",              kMjXRealVec,   0, nullptr,    0},
+
+  //-------------------------------- <fixed> (tendon section) -------------------------------------
+  {"fixed", "name",              kMjXString,    0, nullptr,    0},
+  {"fixed", "class",             kMjXString,    0, nullptr,    0},
+  {"fixed", "group",             kMjXInt,       0, nullptr,    0},
+  {"fixed", "limited",           kMjXEnum,      0, TFAuto_map, 3},
+  {"fixed", "actuatorfrclimited",kMjXEnum,      0, TFAuto_map, 3},
+  {"fixed", "range",             kMjXRealN,     2, nullptr,    0},
+  {"fixed", "actuatorfrcrange",  kMjXRealN,     2, nullptr,    0},
+  {"fixed", "solreflimit",       kMjXRealUpToN, 2, nullptr,    0},
+  {"fixed", "solimplimit",       kMjXRealUpToN, 5, nullptr,    0},
+  {"fixed", "solreffriction",    kMjXRealUpToN, 2, nullptr,    0},
+  {"fixed", "solimpfriction",    kMjXRealUpToN, 5, nullptr,    0},
+  {"fixed", "frictionloss",      kMjXReal,      0, nullptr,    0},
+  {"fixed", "springlength",      kMjXRealUpToN, 2, nullptr,    0},
+  {"fixed", "margin",            kMjXReal,      0, nullptr,    0},
+  {"fixed", "stiffness",         kMjXRealUpToN, 3, nullptr,    0},
+  {"fixed", "damping",           kMjXRealUpToN, 3, nullptr,    0},
+  {"fixed", "armature",          kMjXReal,      0, nullptr,    0},
+  {"fixed", "user",              kMjXRealVec,   0, nullptr,    0},
+
+  //-------------------------------- <spatial> children: site/geom/pulley -------------------------
+  // Inside <spatial>, children are <site site=...>/<geom geom=... sidesite=...>/<pulley divisor=...>
+  // However the child element name "site" collides with top-level <site>, and our lookup is by
+  // (element, attr). We accept the collision: "site" attribute "site" is unique to the wrap,
+  // but "geom" attribute "geom" is also unique.
+  {"pulley", "divisor", kMjXReal,   0, nullptr, 0},
+
+  //-------------------------------- <fixed><joint> (tendon wrap) ---------------------------------
+  // The <joint> child of <fixed> has attrs {joint, coef}. Our top-level <joint> already uses
+  // "name"/"class"/... so the string refs "joint" and real "coef" are distinct here; map them.
+  {"joint", "coef",    kMjXReal,   0, nullptr, 0},
+
+  //-------------------------------- actuator (shared attrs) --------------------------------------
+  // Attributes common to all actuator shortcut types; listed per element so per-(element,attr)
+  // lookups work.
+#define ACT_COMMON(elem) \
+  {elem, "name",          kMjXString,    0, nullptr,    0}, \
+  {elem, "class",         kMjXString,    0, nullptr,    0}, \
+  {elem, "group",         kMjXInt,       0, nullptr,    0}, \
+  {elem, "nsample",       kMjXInt,       0, nullptr,    0}, \
+  {elem, "interp",        kMjXEnum,      0, interp_map, 3}, \
+  {elem, "delay",         kMjXReal,      0, nullptr,    0}, \
+  {elem, "ctrllimited",   kMjXEnum,      0, TFAuto_map, 3}, \
+  {elem, "forcelimited",  kMjXEnum,      0, TFAuto_map, 3}, \
+  {elem, "ctrlrange",     kMjXRealN,     2, nullptr,    0}, \
+  {elem, "forcerange",    kMjXRealN,     2, nullptr,    0}, \
+  {elem, "lengthrange",   kMjXRealN,     2, nullptr,    0}, \
+  {elem, "gear",          kMjXRealUpToN, 6, nullptr,    0}, \
+  {elem, "damping",       kMjXRealUpToN, 3, nullptr,    0}, \
+  {elem, "armature",      kMjXReal,      0, nullptr,    0}, \
+  {elem, "cranklength",   kMjXReal,      0, nullptr,    0}, \
+  {elem, "user",          kMjXRealVec,   0, nullptr,    0}, \
+  {elem, "joint",         kMjXString,    0, nullptr,    0}, \
+  {elem, "jointinparent", kMjXString,    0, nullptr,    0}, \
+  {elem, "tendon",        kMjXString,    0, nullptr,    0}, \
+  {elem, "slidersite",    kMjXString,    0, nullptr,    0}, \
+  {elem, "cranksite",     kMjXString,    0, nullptr,    0}, \
+  {elem, "site",          kMjXString,    0, nullptr,    0}, \
+  {elem, "refsite",       kMjXString,    0, nullptr,    0}, \
+  {elem, "body",          kMjXString,    0, nullptr,    0}
+
+  ACT_COMMON("general"),
+  {"general", "actlimited", kMjXEnum,      0, TFAuto_map, 3},
+  {"general", "actrange",   kMjXRealN,     2, nullptr,    0},
+  {"general", "actdim",     kMjXInt,       0, nullptr,    0},
+  {"general", "dyntype",    kMjXEnum,      0, dyn_map,    7},
+  {"general", "gaintype",   kMjXEnum,      0, gain_map,   5},
+  {"general", "biastype",   kMjXEnum,      0, bias_map,   5},
+  {"general", "dynprm",     kMjXRealUpToN, mjNDYN,  nullptr, 0},
+  {"general", "gainprm",    kMjXRealUpToN, mjNGAIN, nullptr, 0},
+  {"general", "biasprm",    kMjXRealUpToN, mjNBIAS, nullptr, 0},
+  {"general", "actearly",   kMjXEnum,      0, bool_map, 2},
+
+  ACT_COMMON("motor"),
+
+  ACT_COMMON("position"),
+  {"position", "kp",           kMjXReal, 0, nullptr, 0},
+  {"position", "kv",           kMjXReal, 0, nullptr, 0},
+  {"position", "dampratio",    kMjXReal, 0, nullptr, 0},
+  {"position", "timeconst",    kMjXReal, 0, nullptr, 0},
+  {"position", "inheritrange", kMjXReal, 0, nullptr, 0},
+
+  ACT_COMMON("velocity"),
+  {"velocity", "kv", kMjXReal, 0, nullptr, 0},
+
+  ACT_COMMON("intvelocity"),
+  {"intvelocity", "kp",           kMjXReal,  0, nullptr, 0},
+  {"intvelocity", "kv",           kMjXReal,  0, nullptr, 0},
+  {"intvelocity", "dampratio",    kMjXReal,  0, nullptr, 0},
+  {"intvelocity", "actrange",     kMjXRealN, 2, nullptr, 0},
+  {"intvelocity", "inheritrange", kMjXReal,  0, nullptr, 0},
+
+  ACT_COMMON("damper"),
+  {"damper", "kv", kMjXReal, 0, nullptr, 0},
+
+  ACT_COMMON("cylinder"),
+  {"cylinder", "timeconst", kMjXReal,  0, nullptr, 0},
+  {"cylinder", "area",      kMjXReal,  0, nullptr, 0},
+  {"cylinder", "diameter",  kMjXReal,  0, nullptr, 0},
+  {"cylinder", "bias",      kMjXRealN, 3, nullptr, 0},
+
+  ACT_COMMON("muscle"),
+  {"muscle", "timeconst", kMjXRealN, 2, nullptr, 0},
+  {"muscle", "tausmooth", kMjXReal,  0, nullptr, 0},
+  {"muscle", "range",     kMjXRealN, 2, nullptr, 0},
+  {"muscle", "force",     kMjXReal,  0, nullptr, 0},
+  {"muscle", "scale",     kMjXReal,  0, nullptr, 0},
+  {"muscle", "lmin",      kMjXReal,  0, nullptr, 0},
+  {"muscle", "lmax",      kMjXReal,  0, nullptr, 0},
+  {"muscle", "vmax",      kMjXReal,  0, nullptr, 0},
+  {"muscle", "fpmax",     kMjXReal,  0, nullptr, 0},
+  {"muscle", "fvmax",     kMjXReal,  0, nullptr, 0},
+
+  // adhesion doesn't use full ACT_COMMON (no joint/tendon/etc), so list manually.
+  {"adhesion", "name",         kMjXString,    0, nullptr,    0},
+  {"adhesion", "class",        kMjXString,    0, nullptr,    0},
+  {"adhesion", "group",        kMjXInt,       0, nullptr,    0},
+  {"adhesion", "nsample",      kMjXInt,       0, nullptr,    0},
+  {"adhesion", "interp",       kMjXEnum,      0, interp_map, 3},
+  {"adhesion", "delay",        kMjXReal,      0, nullptr,    0},
+  {"adhesion", "forcelimited", kMjXEnum,      0, TFAuto_map, 3},
+  {"adhesion", "ctrlrange",    kMjXRealN,     2, nullptr,    0},
+  {"adhesion", "forcerange",   kMjXRealN,     2, nullptr,    0},
+  {"adhesion", "user",         kMjXRealVec,   0, nullptr,    0},
+  {"adhesion", "body",         kMjXString,    0, nullptr,    0},
+  {"adhesion", "gain",         kMjXReal,      0, nullptr,    0},
+
+  ACT_COMMON("dcmotor"),
+  {"dcmotor", "motorconst", kMjXRealUpToN, 2, nullptr,         0},
+  {"dcmotor", "resistance", kMjXReal,      0, nullptr,         0},
+  {"dcmotor", "nominal",    kMjXRealUpToN, 3, nullptr,         0},
+  {"dcmotor", "saturation", kMjXRealUpToN, 3, nullptr,         0},
+  {"dcmotor", "inductance", kMjXRealUpToN, 2, nullptr,         0},
+  {"dcmotor", "cogging",    kMjXRealUpToN, 3, nullptr,         0},
+  {"dcmotor", "controller", kMjXRealUpToN, 6, nullptr,         0},
+  {"dcmotor", "thermal",    kMjXRealUpToN, 6, nullptr,         0},
+  {"dcmotor", "lugre",      kMjXRealUpToN, 5, nullptr,         0},
+  {"dcmotor", "input",      kMjXEnum,      0, dcmotorinput_map, 3},
+
+#undef ACT_COMMON
+
+  //-------------------------------- <extension> ---------------------------------------------------
+  {"plugin",   "plugin",   kMjXString, 0, nullptr, 0},
+  {"plugin",   "instance", kMjXString, 0, nullptr, 0},
+  {"plugin",   "name",     kMjXString, 0, nullptr, 0},
+  {"plugin",   "class",    kMjXString, 0, nullptr, 0},
+  {"instance", "name",     kMjXString, 0, nullptr, 0},
+  {"config",   "key",      kMjXString, 0, nullptr, 0},
+  {"config",   "value",    kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <custom> ------------------------------------------------------
+  {"numeric", "name", kMjXString,  0, nullptr, 0},
+  {"numeric", "size", kMjXInt,     0, nullptr, 0},
+  {"numeric", "data", kMjXRealVec, 0, nullptr, 0},
+
+  {"text",    "name", kMjXString,  0, nullptr, 0},
+  {"text",    "data", kMjXString,  0, nullptr, 0},
+
+  {"tuple",   "name",    kMjXString, 0, nullptr, 0},
+  {"element", "objtype", kMjXString, 0, nullptr, 0},
+  {"element", "objname", kMjXString, 0, nullptr, 0},
+  {"element", "prm",     kMjXReal,   0, nullptr, 0},
+
+  //-------------------------------- <asset><hfield> -----------------------------------------------
+  {"hfield", "name",         kMjXString,  0, nullptr, 0},
+  {"hfield", "content_type", kMjXString,  0, nullptr, 0},
+  {"hfield", "file",         kMjXString,  0, nullptr, 0},
+  {"hfield", "nrow",         kMjXInt,     0, nullptr, 0},
+  {"hfield", "ncol",         kMjXInt,     0, nullptr, 0},
+  {"hfield", "size",         kMjXRealN,   4, nullptr, 0},
+  {"hfield", "elevation",    kMjXRealVec, 0, nullptr, 0},
+
+  //-------------------------------- <asset><skin> / <deformable><skin> ---------------------------
+  {"skin", "name",     kMjXString,  0, nullptr,  0},
+  {"skin", "file",     kMjXString,  0, nullptr,  0},
+  {"skin", "material", kMjXString,  0, nullptr,  0},
+  {"skin", "rgba",     kMjXRealN,   4, nullptr,  0},
+  {"skin", "inflate",  kMjXReal,    0, nullptr,  0},
+  {"skin", "vertex",   kMjXRealVec, 0, nullptr,  0},
+  {"skin", "texcoord", kMjXRealVec, 0, nullptr,  0},
+  {"skin", "face",     kMjXIntVec,  0, nullptr,  0},
+  {"skin", "group",    kMjXInt,     0, nullptr,  0},
+  // composite's <skin> sub-element
+  {"skin", "texcoord_composite", kMjXEnum, 0, bool_map, 2},  // unused placeholder
+  {"skin", "subgrid",  kMjXInt,     0, nullptr,  0},
+
+  {"bone", "body",       kMjXString,  0, nullptr, 0},
+  {"bone", "bindpos",    kMjXRealN,   3, nullptr, 0},
+  {"bone", "bindquat",   kMjXRealN,   4, nullptr, 0},
+  {"bone", "vertid",     kMjXIntVec,  0, nullptr, 0},
+  {"bone", "vertweight", kMjXRealVec, 0, nullptr, 0},
+
+  //-------------------------------- <asset><texture> ----------------------------------------------
+  {"texture", "name",         kMjXString,  0, nullptr,         0},
+  {"texture", "type",         kMjXEnum,    0, texture_map,     3},
+  {"texture", "colorspace",   kMjXEnum,    0, colorspace_map,  3},
+  {"texture", "content_type", kMjXString,  0, nullptr,         0},
+  {"texture", "file",         kMjXString,  0, nullptr,         0},
+  {"texture", "gridsize",     kMjXIntN,    2, nullptr,         0},
+  {"texture", "gridlayout",   kMjXString,  0, nullptr,         0},
+  {"texture", "fileright",    kMjXString,  0, nullptr,         0},
+  {"texture", "fileleft",     kMjXString,  0, nullptr,         0},
+  {"texture", "fileup",       kMjXString,  0, nullptr,         0},
+  {"texture", "filedown",     kMjXString,  0, nullptr,         0},
+  {"texture", "filefront",    kMjXString,  0, nullptr,         0},
+  {"texture", "fileback",     kMjXString,  0, nullptr,         0},
+  {"texture", "builtin",      kMjXEnum,    0, builtin_map,     4},
+  {"texture", "rgb1",         kMjXRealN,   3, nullptr,         0},
+  {"texture", "rgb2",         kMjXRealN,   3, nullptr,         0},
+  {"texture", "mark",         kMjXEnum,    0, mark_map,        4},
+  {"texture", "markrgb",      kMjXRealN,   3, nullptr,         0},
+  {"texture", "random",       kMjXReal,    0, nullptr,         0},
+  {"texture", "width",        kMjXInt,     0, nullptr,         0},
+  {"texture", "height",       kMjXInt,     0, nullptr,         0},
+  {"texture", "hflip",        kMjXEnum,    0, bool_map,        2},
+  {"texture", "vflip",        kMjXEnum,    0, bool_map,        2},
+  {"texture", "nchannel",     kMjXInt,     0, nullptr,         0},
+
+  //-------------------------------- <asset><model> ------------------------------------------------
+  {"model", "name",         kMjXString, 0, nullptr, 0},
+  {"model", "file",         kMjXString, 0, nullptr, 0},
+  {"model", "content_type", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <body> / <worldbody> -----------------------------------------
+  {"body", "name",       kMjXString,  0, nullptr,       0},
+  {"body", "childclass", kMjXString,  0, nullptr,       0},
+  {"body", "pos",        kMjXRealN,   3, nullptr,       0},
+  {"body", "quat",       kMjXRealN,   4, nullptr,       0},
+  {"body", "mocap",      kMjXEnum,    0, bool_map,      2},
+  {"body", "axisangle",  kMjXRealN,   4, nullptr,       0},
+  {"body", "xyaxes",     kMjXRealN,   6, nullptr,       0},
+  {"body", "zaxis",      kMjXRealN,   3, nullptr,       0},
+  {"body", "euler",      kMjXRealN,   3, nullptr,       0},
+  {"body", "gravcomp",   kMjXReal,    0, nullptr,       0},
+  {"body", "sleep",      kMjXEnum,    0, bodysleep_map, 4},
+  {"body", "user",       kMjXRealVec, 0, nullptr,       0},
+
+  //-------------------------------- <body><inertial> ---------------------------------------------
+  {"inertial", "pos",          kMjXRealN, 3, nullptr, 0},
+  {"inertial", "quat",         kMjXRealN, 4, nullptr, 0},
+  {"inertial", "mass",         kMjXReal,  0, nullptr, 0},
+  {"inertial", "diaginertia",  kMjXRealN, 3, nullptr, 0},
+  {"inertial", "axisangle",    kMjXRealN, 4, nullptr, 0},
+  {"inertial", "xyaxes",       kMjXRealN, 6, nullptr, 0},
+  {"inertial", "zaxis",        kMjXRealN, 3, nullptr, 0},
+  {"inertial", "euler",        kMjXRealN, 3, nullptr, 0},
+  {"inertial", "fullinertia",  kMjXRealN, 6, nullptr, 0},
+
+  //-------------------------------- <body><frame> ------------------------------------------------
+  {"frame", "name",       kMjXString, 0, nullptr, 0},
+  {"frame", "childclass", kMjXString, 0, nullptr, 0},
+  {"frame", "pos",        kMjXRealN,  3, nullptr, 0},
+  {"frame", "quat",       kMjXRealN,  4, nullptr, 0},
+  {"frame", "axisangle",  kMjXRealN,  4, nullptr, 0},
+  {"frame", "xyaxes",     kMjXRealN,  6, nullptr, 0},
+  {"frame", "zaxis",      kMjXRealN,  3, nullptr, 0},
+  {"frame", "euler",      kMjXRealN,  3, nullptr, 0},
+
+  //-------------------------------- <body><replicate> --------------------------------------------
+  {"replicate", "count",      kMjXInt,    0, nullptr, 0},
+  {"replicate", "offset",     kMjXRealN,  3, nullptr, 0},
+  {"replicate", "euler",      kMjXRealN,  3, nullptr, 0},
+  {"replicate", "sep",        kMjXString, 0, nullptr, 0},
+  {"replicate", "childclass", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <body><attach> -----------------------------------------------
+  {"attach", "model",  kMjXString, 0, nullptr, 0},
+  {"attach", "body",   kMjXString, 0, nullptr, 0},
+  {"attach", "prefix", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <composite> --------------------------------------------------
+  {"composite", "prefix",  kMjXString,  0, nullptr,  0},
+  {"composite", "type",    kMjXEnum,    0, comp_map, mjNCOMPTYPES},
+  {"composite", "count",   kMjXIntVec,  0, nullptr,  0},
+  {"composite", "offset",  kMjXRealN,   3, nullptr,  0},
+  {"composite", "vertex",  kMjXRealVec, 0, nullptr,  0},
+  {"composite", "initial", kMjXString,  0, nullptr,  0},
+  {"composite", "curve",   kMjXString,  0, nullptr,  0},
+  {"composite", "size",    kMjXRealN,   3, nullptr,  0},
+  {"composite", "quat",    kMjXRealN,   4, nullptr,  0},
+
+  //-------------------------------- <flexcomp> ---------------------------------------------------
+  {"flexcomp", "name",       kMjXString,  0, nullptr,   0},
+  {"flexcomp", "type",       kMjXEnum,    0, fcomp_map, mjNFCOMPTYPES},
+  {"flexcomp", "group",      kMjXInt,     0, nullptr,   0},
+  {"flexcomp", "dim",        kMjXInt,     0, nullptr,   0},
+  {"flexcomp", "dof",        kMjXEnum,    0, fdof_map,  mjNFCOMPDOFS},
+  {"flexcomp", "count",      kMjXIntN,    3, nullptr,   0},
+  {"flexcomp", "cellcount",  kMjXIntN,    3, nullptr,   0},
+  {"flexcomp", "spacing",    kMjXRealN,   3, nullptr,   0},
+  {"flexcomp", "radius",     kMjXReal,    0, nullptr,   0},
+  {"flexcomp", "rigid",      kMjXEnum,    0, bool_map,  2},
+  {"flexcomp", "mass",       kMjXReal,    0, nullptr,   0},
+  {"flexcomp", "inertiabox", kMjXReal,    0, nullptr,   0},
+  {"flexcomp", "scale",      kMjXRealN,   3, nullptr,   0},
+  {"flexcomp", "file",       kMjXString,  0, nullptr,   0},
+  {"flexcomp", "point",      kMjXRealVec, 0, nullptr,   0},
+  {"flexcomp", "element",    kMjXIntVec,  0, nullptr,   0},
+  {"flexcomp", "texcoord",   kMjXRealVec, 0, nullptr,   0},
+  {"flexcomp", "material",   kMjXString,  0, nullptr,   0},
+  {"flexcomp", "rgba",       kMjXRealN,   4, nullptr,   0},
+  {"flexcomp", "flatskin",   kMjXEnum,    0, bool_map,  2},
+  {"flexcomp", "pos",        kMjXRealN,   3, nullptr,   0},
+  {"flexcomp", "quat",       kMjXRealN,   4, nullptr,   0},
+  {"flexcomp", "axisangle",  kMjXRealN,   4, nullptr,   0},
+  {"flexcomp", "xyaxes",     kMjXRealN,   6, nullptr,   0},
+  {"flexcomp", "zaxis",      kMjXRealN,   3, nullptr,   0},
+  {"flexcomp", "euler",      kMjXRealN,   3, nullptr,   0},
+  {"flexcomp", "origin",     kMjXRealN,   3, nullptr,   0},
+
+  //-------------------------------- <flexcomp>/<flex> children -----------------------------------
+  // <edge> under <flex> has {stiffness, damping}; under <flexcomp> adds {equality, solref, solimp}.
+  // We include the union.
+  {"edge", "equality",  kMjXEnum,      0, flexeq_map, 4},
+  {"edge", "solref",    kMjXRealUpToN, 2, nullptr,  0},
+  {"edge", "solimp",    kMjXRealUpToN, 5, nullptr,  0},
+  {"edge", "stiffness", kMjXReal,      0, nullptr,  0},
+  {"edge", "damping",   kMjXReal,      0, nullptr,  0},
+
+  {"elasticity", "young",     kMjXReal,      0, nullptr,         0},
+  {"elasticity", "poisson",   kMjXReal,      0, nullptr,         0},
+  {"elasticity", "damping",   kMjXReal,      0, nullptr,         0},
+  {"elasticity", "thickness", kMjXReal,      0, nullptr,         0},
+  {"elasticity", "elastic2d", kMjXEnum,      0, elastic2d_map,   5},
+
+  {"contact", "contype",      kMjXInt,       0, nullptr,      0},
+  {"contact", "conaffinity",  kMjXInt,       0, nullptr,      0},
+  {"contact", "condim",       kMjXInt,       0, nullptr,      0},
+  {"contact", "priority",     kMjXInt,       0, nullptr,      0},
+  {"contact", "friction",     kMjXRealUpToN, 3, nullptr,      0},
+  {"contact", "solmix",       kMjXReal,      0, nullptr,      0},
+  {"contact", "solref",       kMjXRealUpToN, 2, nullptr,      0},
+  {"contact", "solimp",       kMjXRealUpToN, 5, nullptr,      0},
+  {"contact", "margin",       kMjXReal,      0, nullptr,      0},
+  {"contact", "gap",          kMjXReal,      0, nullptr,      0},
+  {"contact", "internal",     kMjXEnum,      0, bool_map,     2},
+  {"contact", "selfcollide",  kMjXEnum,      0, flexself_map, 5},
+  {"contact", "activelayers", kMjXInt,       0, nullptr,      0},
+  {"contact", "passive",      kMjXEnum,      0, bool_map,     2},
+
+  {"pin", "id",        kMjXIntVec,   0, nullptr, 0},
+  {"pin", "range",     kMjXRealN,    2, nullptr, 0},
+  {"pin", "grid",      kMjXIntVec,   0, nullptr, 0},
+  {"pin", "gridrange", kMjXIntVec,   0, nullptr, 0},
+
+  //-------------------------------- <composite> children -----------------------------------------
+  // Most of composite's children re-use element names that already have entries.
+  // "kind" is specific to composite's <joint>; treat as a composite-joint attr via shared key.
+
+  //-------------------------------- <deformable><flex> -------------------------------------------
+  {"flex", "name",         kMjXString,   0, nullptr,   0},
+  {"flex", "group",        kMjXInt,      0, nullptr,   0},
+  {"flex", "dim",          kMjXInt,      0, nullptr,   0},
+  {"flex", "radius",       kMjXReal,     0, nullptr,   0},
+  {"flex", "material",     kMjXString,   0, nullptr,   0},
+  {"flex", "rgba",         kMjXRealN,    4, nullptr,   0},
+  {"flex", "flatskin",     kMjXEnum,     0, bool_map,  2},
+  {"flex", "body",         kMjXString,   0, nullptr,   0},
+  {"flex", "vertex",       kMjXRealVec,  0, nullptr,   0},
+  {"flex", "element",      kMjXIntVec,   0, nullptr,   0},
+  {"flex", "texcoord",     kMjXRealVec,  0, nullptr,   0},
+  {"flex", "elemtexcoord", kMjXIntVec,   0, nullptr,   0},
+  {"flex", "node",         kMjXString,   0, nullptr,   0},
+  {"flex", "cellcount",    kMjXIntN,     3, nullptr,   0},
+  {"flex", "dof",          kMjXEnum,     0, fdof_map,  mjNFCOMPDOFS},
+  // equality-flex refs (used in <equality><flex>)
+  {"flex", "class",        kMjXString,   0, nullptr,   0},
+  {"flex", "flex",         kMjXString,   0, nullptr,   0},
+  {"flex", "active",       kMjXEnum,     0, bool_map,  2},
+  {"flex", "solref",       kMjXRealUpToN,2, nullptr,   0},
+  {"flex", "solimp",       kMjXRealUpToN,5, nullptr,   0},
+
+  //-------------------------------- <contact> ----------------------------------------------------
+  {"exclude", "name",  kMjXString, 0, nullptr, 0},
+  {"exclude", "body1", kMjXString, 0, nullptr, 0},
+  {"exclude", "body2", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <equality> children ------------------------------------------
+  {"connect", "name",     kMjXString,    0, nullptr,  0},
+  {"connect", "class",    kMjXString,    0, nullptr,  0},
+  {"connect", "body1",    kMjXString,    0, nullptr,  0},
+  {"connect", "body2",    kMjXString,    0, nullptr,  0},
+  {"connect", "anchor",   kMjXRealN,     3, nullptr,  0},
+  {"connect", "site1",    kMjXString,    0, nullptr,  0},
+  {"connect", "site2",    kMjXString,    0, nullptr,  0},
+  {"connect", "active",   kMjXEnum,      0, bool_map, 2},
+  {"connect", "solref",   kMjXRealUpToN, 2, nullptr,  0},
+  {"connect", "solimp",   kMjXRealUpToN, 5, nullptr,  0},
+
+  {"weld", "name",        kMjXString,    0, nullptr,  0},
+  {"weld", "class",       kMjXString,    0, nullptr,  0},
+  {"weld", "body1",       kMjXString,    0, nullptr,  0},
+  {"weld", "body2",       kMjXString,    0, nullptr,  0},
+  {"weld", "relpose",     kMjXRealN,     7, nullptr,  0},
+  {"weld", "anchor",      kMjXRealN,     3, nullptr,  0},
+  {"weld", "site1",       kMjXString,    0, nullptr,  0},
+  {"weld", "site2",       kMjXString,    0, nullptr,  0},
+  {"weld", "active",      kMjXEnum,      0, bool_map, 2},
+  {"weld", "solref",      kMjXRealUpToN, 2, nullptr,  0},
+  {"weld", "solimp",      kMjXRealUpToN, 5, nullptr,  0},
+  {"weld", "torquescale", kMjXReal,      0, nullptr,  0},
+
+  // <equality><joint> has a distinct set of attrs from top-level <joint>.
+  // Since our lookup key is (element, attr), we add the ones not already covered
+  // by top-level <joint> (joint1, joint2, polycoef).
+  {"joint", "joint1",   kMjXString,    0, nullptr,  0},
+  {"joint", "joint2",   kMjXString,    0, nullptr,  0},
+  {"joint", "polycoef", kMjXRealUpToN, 5, nullptr,  0},
+  {"joint", "active",   kMjXEnum,      0, bool_map, 2},
+  {"joint", "solref",   kMjXRealUpToN, 2, nullptr,  0},
+  {"joint", "solimp",   kMjXRealUpToN, 5, nullptr,  0},
+  {"joint", "kind",     kMjXString,    0, nullptr,  0},  // composite joint kind
+  {"joint", "joint",    kMjXString,    0, nullptr,  0},  // tendon wrap
+
+  // <equality><tendon>
+  {"tendon", "name",     kMjXString,    0, nullptr,  0},
+  {"tendon", "class",    kMjXString,    0, nullptr,  0},
+  {"tendon", "tendon1",  kMjXString,    0, nullptr,  0},
+  {"tendon", "tendon2",  kMjXString,    0, nullptr,  0},
+  {"tendon", "polycoef", kMjXRealUpToN, 5, nullptr,  0},
+
+  // flexvert / flexstrain equality variants
+  {"flexvert", "name",   kMjXString,    0, nullptr,  0},
+  {"flexvert", "class",  kMjXString,    0, nullptr,  0},
+  {"flexvert", "flex",   kMjXString,    0, nullptr,  0},
+  {"flexvert", "active", kMjXEnum,      0, bool_map, 2},
+  {"flexvert", "solref", kMjXRealUpToN, 2, nullptr,  0},
+  {"flexvert", "solimp", kMjXRealUpToN, 5, nullptr,  0},
+
+  {"flexstrain", "name",   kMjXString,    0, nullptr,  0},
+  {"flexstrain", "class",  kMjXString,    0, nullptr,  0},
+  {"flexstrain", "flex",   kMjXString,    0, nullptr,  0},
+  {"flexstrain", "active", kMjXEnum,      0, bool_map, 2},
+  {"flexstrain", "solref", kMjXRealUpToN, 2, nullptr,  0},
+  {"flexstrain", "solimp", kMjXRealUpToN, 5, nullptr,  0},
+
+  //-------------------------------- <tendon> wrap subelements ------------------------------------
+  // <site site="..."/>, <geom geom="..." sidesite="..."/>
+  {"site", "site",     kMjXString, 0, nullptr, 0},
+  {"geom", "geom",     kMjXString, 0, nullptr, 0},
+  {"geom", "sidesite", kMjXString, 0, nullptr, 0},
+
+  //-------------------------------- <sensor> ------------------------------------------------------
+  // shared sensor attrs
+#define SENS_COMMON(elem) \
+  {elem, "name",     kMjXString,    0, nullptr,    0}, \
+  {elem, "nsample",  kMjXInt,       0, nullptr,    0}, \
+  {elem, "interp",   kMjXEnum,      0, interp_map, 3}, \
+  {elem, "delay",    kMjXReal,      0, nullptr,    0}, \
+  {elem, "interval", kMjXRealUpToN, 2, nullptr,    0}, \
+  {elem, "cutoff",   kMjXReal,      0, nullptr,    0}, \
+  {elem, "noise",    kMjXReal,      0, nullptr,    0}, \
+  {elem, "user",     kMjXRealVec,   0, nullptr,    0}
+
+  SENS_COMMON("touch"),         {"touch",         "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("accelerometer"), {"accelerometer", "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("velocimeter"),   {"velocimeter",   "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("gyro"),          {"gyro",          "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("force"),         {"force",         "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("torque"),        {"torque",        "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("magnetometer"),  {"magnetometer",  "site",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("camprojection"), {"camprojection", "site",     kMjXString, 0, nullptr, 0},
+                                {"camprojection", "camera",   kMjXString, 0, nullptr, 0},
+  SENS_COMMON("rangefinder"),   {"rangefinder",   "site",     kMjXString, 0, nullptr, 0},
+                                {"rangefinder",   "camera",   kMjXString, 0, nullptr, 0},
+                                {"rangefinder",   "data",     kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointpos"),      {"jointpos",      "joint",    kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointvel"),      {"jointvel",      "joint",    kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonpos"),     {"tendonpos",     "tendon",   kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonvel"),     {"tendonvel",     "tendon",   kMjXString, 0, nullptr, 0},
+  SENS_COMMON("actuatorpos"),   {"actuatorpos",   "actuator", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("actuatorvel"),   {"actuatorvel",   "actuator", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("actuatorfrc"),   {"actuatorfrc",   "actuator", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointactuatorfrc"),  {"jointactuatorfrc",  "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonactuatorfrc"), {"tendonactuatorfrc", "tendon", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("ballquat"),          {"ballquat",          "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("ballangvel"),        {"ballangvel",        "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointlimitpos"),     {"jointlimitpos",     "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointlimitvel"),     {"jointlimitvel",     "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("jointlimitfrc"),     {"jointlimitfrc",     "joint",  kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonlimitpos"),    {"tendonlimitpos",    "tendon", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonlimitvel"),    {"tendonlimitvel",    "tendon", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("tendonlimitfrc"),    {"tendonlimitfrc",    "tendon", kMjXString, 0, nullptr, 0},
+
+#define SENS_FRAME(elem) \
+  SENS_COMMON(elem), \
+  {elem, "objtype", kMjXString, 0, nullptr, 0}, \
+  {elem, "objname", kMjXString, 0, nullptr, 0}, \
+  {elem, "reftype", kMjXString, 0, nullptr, 0}, \
+  {elem, "refname", kMjXString, 0, nullptr, 0}
+
+  SENS_FRAME("framepos"),
+  SENS_FRAME("framequat"),
+  SENS_FRAME("framexaxis"),
+  SENS_FRAME("frameyaxis"),
+  SENS_FRAME("framezaxis"),
+  SENS_FRAME("framelinvel"),
+  SENS_FRAME("frameangvel"),
+  SENS_COMMON("framelinacc"), {"framelinacc", "objtype", kMjXString, 0, nullptr, 0},
+                              {"framelinacc", "objname", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("frameangacc"), {"frameangacc", "objtype", kMjXString, 0, nullptr, 0},
+                              {"frameangacc", "objname", kMjXString, 0, nullptr, 0},
+
+  SENS_COMMON("subtreecom"),    {"subtreecom",    "body", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("subtreelinvel"), {"subtreelinvel", "body", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("subtreeangmom"), {"subtreeangmom", "body", kMjXString, 0, nullptr, 0},
+
+  SENS_COMMON("insidesite"),
+  {"insidesite", "site",    kMjXString, 0, nullptr, 0},
+  {"insidesite", "objtype", kMjXString, 0, nullptr, 0},
+  {"insidesite", "objname", kMjXString, 0, nullptr, 0},
+
+  // distance/normal/fromto share the same attrs
+  SENS_COMMON("distance"),
+  {"distance", "geom1", kMjXString, 0, nullptr, 0},
+  {"distance", "geom2", kMjXString, 0, nullptr, 0},
+  {"distance", "body1", kMjXString, 0, nullptr, 0},
+  {"distance", "body2", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("normal"),
+  {"normal", "geom1", kMjXString, 0, nullptr, 0},
+  {"normal", "geom2", kMjXString, 0, nullptr, 0},
+  {"normal", "body1", kMjXString, 0, nullptr, 0},
+  {"normal", "body2", kMjXString, 0, nullptr, 0},
+  SENS_COMMON("fromto"),
+  {"fromto", "geom1", kMjXString, 0, nullptr, 0},
+  {"fromto", "geom2", kMjXString, 0, nullptr, 0},
+  {"fromto", "body1", kMjXString, 0, nullptr, 0},
+  {"fromto", "body2", kMjXString, 0, nullptr, 0},
+
+  // contact sensor
+  {"contact", "name",     kMjXString, 0, nullptr,    0},
+  {"contact", "geom1",    kMjXString, 0, nullptr,    0},
+  {"contact", "geom2",    kMjXString, 0, nullptr,    0},
+  {"contact", "body1",    kMjXString, 0, nullptr,    0},
+  {"contact", "body2",    kMjXString, 0, nullptr,    0},
+  {"contact", "subtree1", kMjXString, 0, nullptr,    0},
+  {"contact", "subtree2", kMjXString, 0, nullptr,    0},
+  {"contact", "site",     kMjXString, 0, nullptr,    0},
+  {"contact", "num",      kMjXInt,    0, nullptr,    0},
+  {"contact", "data",     kMjXString, 0, nullptr,    0},  // space-separated enum bitflags
+  {"contact", "reduce",   kMjXEnum,   0, reduce_map, 4},
+  {"contact", "nsample",  kMjXInt,    0, nullptr,    0},
+  {"contact", "interp",   kMjXEnum,   0, interp_map, 3},
+  {"contact", "delay",    kMjXReal,   0, nullptr,    0},
+  {"contact", "interval", kMjXRealUpToN, 2, nullptr, 0},
+  {"contact", "cutoff",   kMjXReal,   0, nullptr,    0},
+  {"contact", "noise",    kMjXReal,   0, nullptr,    0},
+  {"contact", "user",     kMjXRealVec,0, nullptr,    0},
+
+  SENS_COMMON("e_potential"),
+  SENS_COMMON("e_kinetic"),
+  SENS_COMMON("clock"),
+
+  {"tactile", "name",    kMjXString,  0, nullptr,    0},
+  {"tactile", "geom",    kMjXString,  0, nullptr,    0},
+  {"tactile", "mesh",    kMjXString,  0, nullptr,    0},
+  {"tactile", "nsample", kMjXInt,     0, nullptr,    0},
+  {"tactile", "interp",  kMjXEnum,    0, interp_map, 3},
+  {"tactile", "delay",   kMjXReal,    0, nullptr,    0},
+  {"tactile", "interval",kMjXRealUpToN, 2, nullptr,  0},
+  {"tactile", "user",    kMjXRealVec, 0, nullptr,    0},
+
+  {"user", "name",      kMjXString,  0, nullptr,       0},
+  {"user", "objtype",   kMjXString,  0, nullptr,       0},
+  {"user", "objname",   kMjXString,  0, nullptr,       0},
+  {"user", "datatype",  kMjXEnum,    0, datatype_map,  4},
+  {"user", "needstage", kMjXEnum,    0, stage_map,     4},
+  {"user", "dim",       kMjXInt,     0, nullptr,       0},
+  {"user", "cutoff",    kMjXReal,    0, nullptr,       0},
+  {"user", "noise",     kMjXReal,    0, nullptr,       0},
+  {"user", "user",      kMjXRealVec, 0, nullptr,       0},
+
+#undef SENS_COMMON
+#undef SENS_FRAME
+
+  //-------------------------------- <keyframe> ---------------------------------------------------
+  {"key", "name",  kMjXString,  0, nullptr, 0},
+  {"key", "time",  kMjXReal,    0, nullptr, 0},
+  {"key", "qpos",  kMjXRealVec, 0, nullptr, 0},
+  {"key", "qvel",  kMjXRealVec, 0, nullptr, 0},
+  {"key", "act",   kMjXRealVec, 0, nullptr, 0},
+  {"key", "mpos",  kMjXRealVec, 0, nullptr, 0},
+  {"key", "mquat", kMjXRealVec, 0, nullptr, 0},
+  {"key", "ctrl",  kMjXRealVec, 0, nullptr, 0},
+};
+// clang-format on
+
+const int kMjXAttrTableSize =
+    sizeof(kMjXAttrTable) / sizeof(kMjXAttrTable[0]);
 
 
 //---------------------------------- class mjXReader implementation --------------------------------
